refactor(tanstack-query-react-plugin): move `maybeResolveImport(...)` into separate file
