+++ packages/tanstack-query-react-plugin/src/ts-factory/getIndexFactory.ts
@@ -1,5 +1,6 @@
 import type { ServiceImportsFactoryOptions } from './getServiceFactory.js';
 import ts from 'typescript';
+import { maybeResolveImport } from '../lib/maybeResolveImport.js';
 
 export const getIndexFactory = ({
   openapiTypesImportPath,
@@ -71,23 +72,3 @@ export const getIndexFactory = ({
     ),
   ];
 };
-
-/**
- * Prepend the import path if it's a relative path.
- * Normally,`openapiTypesImportPath` is a relative path to the services' directory.
- */
-export function maybeResolveImport<T extends string | undefined>({
-  openapiTypesImportPath,
-  servicesDirName,
-}: {
-  openapiTypesImportPath: T;
-  servicesDirName: string;
-}) {
-  if (openapiTypesImportPath?.startsWith('../')) {
-    return `./${openapiTypesImportPath.slice(3)}`;
-  } else if (openapiTypesImportPath?.startsWith('./')) {
-    return `./${servicesDirName}/${openapiTypesImportPath.slice(2)}`;
-  }
-
-  return openapiTypesImportPath;
-}
